board 3.16

+++ src/main/java/com/onestop/GJ/board/notice/controller/BoardNoticeControllerImpl.java
@@ -1,7 +1,9 @@
 package com.onestop.GJ.board.notice.controller;
 
 import java.io.File;
+import java.io.PrintWriter;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -31,8 +33,6 @@ import com.onestop.GJ.board.notice.vo.BoardNoticeImageVO;
 import com.onestop.GJ.board.notice.vo.BoardNoticeVO;
 import com.onestop.GJ.member.vo.MemberVO;
 
-import net.sf.json.JSONObject;
-
 @Controller("boardController")
 public class BoardNoticeControllerImpl implements BoardNoticeController {
 	private static String ARTICLE_IMAGE_REPO = "C:\\GJ\\file_repo\\board\\notice";
@@ -44,54 +44,54 @@ public class BoardNoticeControllerImpl implements BoardNoticeController {
 	@Override
 	@RequestMapping(value = { "/boardNotice/listArticles.do" }, method = { RequestMethod.GET, RequestMethod.POST })
 	public ModelAndView listArticles(HttpServletRequest request, HttpServletResponse response) throws Exception {
-		
-		String _section=request.getParameter("section");
-		String _pageNum=request.getParameter("pageNum");
-		int section = Integer.parseInt(((_section==null)? "1":_section));
-		int pageNum= Integer.parseInt(((_pageNum==null)? "1":_pageNum));
-		Map pagingMap=new HashMap();
-		pagingMap.put("section",  section);
-		pagingMap.put("pageNum",  pageNum);
+
+		String _section = request.getParameter("section");
+		String _pageNum = request.getParameter("pageNum");
+		int section = Integer.parseInt(((_section == null) ? "1" : _section));
+		int pageNum = Integer.parseInt(((_pageNum == null) ? "1" : _pageNum));
+		Map pagingMap = new HashMap();
+		pagingMap.put("section", section);
+		pagingMap.put("pageNum", pageNum);
 		Map articlesMap = boardService.listArticles(pagingMap);
-		
+
 		articlesMap.put("section", section);
 		articlesMap.put("pageNum", pageNum);
-		
+
 		String viewName = (String) request.getAttribute("viewName");
 		ModelAndView mav = new ModelAndView(viewName);
 		mav.addObject("articlesMap", articlesMap);
 		return mav;
-		
+
 	}
-	
-	//검색창
+
+	// 검색창
 	@Override
-	@RequestMapping(value="/boardNotice/searchBoardList.do" ,method = RequestMethod.GET)
-	public ModelAndView searchBoardList(@RequestParam("searchWord") String searchWord,
-			                       HttpServletRequest request, HttpServletResponse response) throws Exception{
-		
-		String _section=request.getParameter("section");
-		String _pageNum=request.getParameter("pageNum");
-		int section = Integer.parseInt(((_section==null)? "1":_section));
-		int pageNum= Integer.parseInt(((_pageNum==null)? "1":_pageNum));
-		Map pagingMap=new HashMap();
-		pagingMap.put("section",  section);
-		pagingMap.put("pageNum",  pageNum);
-		pagingMap.put("searchWord",  searchWord);
+	@RequestMapping(value = "/boardNotice/searchBoardList.do", method = RequestMethod.GET)
+	public ModelAndView searchBoardList(@RequestParam("searchWord") String searchWord, HttpServletRequest request,
+			HttpServletResponse response) throws Exception {
+		response.setContentType("text/html;charset=utf-8");
+		response.setCharacterEncoding("utf-8");
+
+		String _section = request.getParameter("section");
+		String _pageNum = request.getParameter("pageNum");
+		int section = Integer.parseInt(((_section == null) ? "1" : _section));
+		int pageNum = Integer.parseInt(((_pageNum == null) ? "1" : _pageNum));
+		Map pagingMap = new HashMap();
+		pagingMap.put("section", section);
+		pagingMap.put("pageNum", pageNum);
+		pagingMap.put("searchWord", searchWord);
 		Map articlesMap = boardService.searchBoardList(pagingMap);
-		
+
 		articlesMap.put("section", section);
 		articlesMap.put("pageNum", pageNum);
 		articlesMap.put("searchWord", searchWord);
-		
-		String viewName=(String)request.getAttribute("viewName");
+
+		String viewName = (String) request.getAttribute("viewName");
 		ModelAndView mav = new ModelAndView(viewName);
 		mav.addObject("articlesMap", articlesMap);
-		System.out.println(articlesMap.size());
+		System.out.println("??" + articlesMap.size());
 		return mav;
-		
 	}
-	
 
 	// 다중 이미지 글쓰기
 	@Override
@@ -214,10 +214,12 @@ public class BoardNoticeControllerImpl implements BoardNoticeController {
 
 //	게시글 상세
 	@RequestMapping(value = "/boardNotice/viewArticle.do", method = RequestMethod.GET)
-	public ModelAndView viewArticle(@RequestParam("noti_NO") int noti_NO, HttpServletRequest request,
-			HttpServletResponse response) throws Exception {
+	public ModelAndView viewArticle(@RequestParam("noti_NO") int noti_NO,
+			@RequestParam(value="removeCompleted", required=false) String removeCompleted,
+			HttpServletRequest request, HttpServletResponse response) throws Exception {
 		String viewName = (String) request.getAttribute("viewName");
 		Map articleMap = boardService.viewArticle(noti_NO);
+		articleMap.put("removeCompleted", removeCompleted );
 		ModelAndView mav = new ModelAndView();
 		mav.setViewName(viewName);
 		mav.addObject("articleMap", articleMap);
@@ -262,25 +264,45 @@ public class BoardNoticeControllerImpl implements BoardNoticeController {
 	public ResponseEntity modArticle(MultipartHttpServletRequest multipartRequest, HttpServletResponse response)
 			throws Exception {
 		multipartRequest.setCharacterEncoding("utf-8");
-		String up_fileName = null;
-
 		Map<String, Object> articleMap = new HashMap<String, Object>();
 		Enumeration enu = multipartRequest.getParameterNames();
 		while (enu.hasMoreElements()) {
 			String name = (String) enu.nextElement();
-			String value = multipartRequest.getParameter(name);
-			articleMap.put(name, value);
+
+			if (name.equals("up_fileNO")) {
+				String[] values = multipartRequest.getParameterValues(name);
+				articleMap.put(name, values);
+			} else if (name.equals("oldFileName")) {
+				String[] values = multipartRequest.getParameterValues(name);
+				articleMap.put(name, values);
+			} else {
+				String value = multipartRequest.getParameter(name);
+				articleMap.put(name, value);
+			}
 		}
 
-		List<String> fileList = upload(multipartRequest);
+		List<String> fileList = uploadModImageFile(multipartRequest);
+		int added_img_num = Integer.parseInt((String) articleMap.get("added_img_num"));
+		int pre_img_num = Integer.parseInt((String) articleMap.get("pre_img_num"));
+
 		List<BoardNoticeImageVO> imageFileList = new ArrayList<BoardNoticeImageVO>();
+		List<BoardNoticeImageVO> modAddimageFileList = new ArrayList<BoardNoticeImageVO>();
 		if (fileList != null && fileList.size() != 0) {
-			for (String fileName : fileList) {
+			String[] up_fileNO = (String[]) articleMap.get("up_fileNO");
+			for (int i = 0; i < added_img_num; i++) {
+				String up_fileName = fileList.get(i);
 				BoardNoticeImageVO boardNoticeImageVO = new BoardNoticeImageVO();
-				boardNoticeImageVO.setUp_fileName(fileName);
-				imageFileList.add(boardNoticeImageVO);
+				if (i < pre_img_num) {
+					boardNoticeImageVO.setUp_fileName(up_fileName);
+					boardNoticeImageVO.setUp_fileNO(Integer.parseInt(up_fileNO[i]));
+					imageFileList.add(boardNoticeImageVO);
+					articleMap.put("imageFileList", imageFileList);
+				} else {
+					boardNoticeImageVO.setUp_fileName(up_fileName);
+					modAddimageFileList.add(boardNoticeImageVO);
+					articleMap.put("modAddimageFileList", modAddimageFileList);
+				}
 			}
-			articleMap.put("imageFileList", imageFileList);
 		}
 
 		String noti_NO = (String) articleMap.get("noti_NO");
@@ -288,22 +310,36 @@ public class BoardNoticeControllerImpl implements BoardNoticeController {
 		ResponseEntity resEnt = null;
 		HttpHeaders responseHeaders = new HttpHeaders();
 		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+		 Collection<Object> value = articleMap.values();
+	      System.out.println("검거" +value);
 		try {
 			boardService.modArticle(articleMap);
 			if (imageFileList != null && imageFileList.size() != 0) {
-				for (BoardNoticeImageVO boardNoticeImageVO : imageFileList) {
-					up_fileName = boardNoticeImageVO.getUp_fileName();
-					File srcFile = new File(ARTICLE_IMAGE_REPO + "\\" + "temp" + "\\" + up_fileName);
-					File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + noti_NO);
-					//destDir = 파일 생성
-					FileUtils.moveFileToDirectory(srcFile, destDir, true);
+				for (int i = 0; i < fileList.size(); i++) {
+					String up_fileName = fileList.get(i);
 
+					if (i < pre_img_num) {
+						if (up_fileName != null) {
+							File srcFile = new File(ARTICLE_IMAGE_REPO + "\\" + "temp" + "\\" + up_fileName);
+							File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + noti_NO);
+							FileUtils.moveFileToDirectory(srcFile, destDir, true);
+
+							String[] oldName = (String[]) articleMap.get("oldFileName");
+							String oldFileName = oldName[i];
+
+							File oldFile = new File(ARTICLE_IMAGE_REPO + "\\" + noti_NO + "\\" + oldFileName);
+							oldFile.delete();
+						}
+					} else {
+						if (up_fileName != null) {
+							File srcFile = new File(ARTICLE_IMAGE_REPO + "\\" + "temp" + "\\" + up_fileName);
+							File destDir = new File(ARTICLE_IMAGE_REPO + "\\" + noti_NO);
+							FileUtils.moveFileToDirectory(srcFile, destDir, true);
+						}
+					}
 				}
-				String originalFileName = (String) articleMap.get("originalFileName");
-				File oldFile = new File(ARTICLE_IMAGE_REPO + "\\" + noti_NO + "\\" + originalFileName);
-				oldFile.delete();
-
 			}
+
 			message = "<script>";
 			message += " alert('글을 수정했습니다.');";
 			message += " location.href='" + multipartRequest.getContextPath() + "/boardNotice/viewArticle.do?noti_NO="
@@ -311,8 +347,14 @@ public class BoardNoticeControllerImpl implements BoardNoticeController {
 			message += " </script>";
 			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
 		} catch (Exception e) {
-			File srcFile = new File(ARTICLE_IMAGE_REPO + "\\" + "temp" + "\\" + up_fileName);
-			srcFile.delete();
+			if (fileList != null && fileList.size() != 0) { // 오류 발생 시 temp 폴더에 업로드된 이미지 파일들을 삭제한다.
+				for (int i = 0; i < fileList.size(); i++) {
+					File srcFile = new File(ARTICLE_IMAGE_REPO + "\\" + "temp" + "\\" + fileList.get(i));
+					srcFile.delete();
+				}
+				e.printStackTrace();
+			}
+
 			message = "<script>";
 			message += " alert('오류가 발생했습니다.다시 수정해주세요');";
 			message += " location.href='" + multipartRequest.getContextPath() + "/boardNotice/viewArticle.do?noti_NO="
@@ -322,30 +364,62 @@ public class BoardNoticeControllerImpl implements BoardNoticeController {
 		}
 		return resEnt;
 	}
-	
-	//키워드
-//	@RequestMapping(value="/keywordSearch.do",method = RequestMethod.GET,produces = "application/text; charset=utf8")
-//	public @ResponseBody String  keywordSearch(@RequestParam("keyword") String keyword,
-//			                                  HttpServletRequest request, HttpServletResponse response) throws Exception{
-//		response.setContentType("text/html;charset=utf-8");
-//		response.setCharacterEncoding("utf-8");
-//		//System.out.println(keyword);
-//		if(keyword == null || keyword.equals(""))
-//		   return null ;
-//	
-//		keyword = keyword.toUpperCase();
-//	    List<String> keywordList =goodsService.keywordSearch(keyword);
-//	    
-//	 // 최종 완성될 JSONObject 선언(전체)
-//		JSONObject jsonObject = new JSONObject();
-//		jsonObject.put("keyword", keywordList);
-//		 		
-//	    String jsonInfo = jsonObject.toString();
-//	   // System.out.println(jsonInfo);
-//	    return jsonInfo ;
-//	}
-	
 
+	// 수정하기에서 이미지 삭제 기능
+		@RequestMapping(value = "/boardNotice/removeModImage.do", method = RequestMethod.POST)
+		@ResponseBody
+		public void removeModImage(HttpServletRequest request, HttpServletResponse response) throws Exception {
 
+			request.setCharacterEncoding("utf-8");
+			response.setContentType("text/html; charset=utf-8");
+			PrintWriter writer = response.getWriter();
+
+			try {
+				String up_fileNO = (String) request.getParameter("up_fileNO");
+				String up_fileName = (String) request.getParameter("up_fileName");
+				String noti_NO = (String) request.getParameter("noti_NO");
+				
+				System.out.println("up_fileNO = " + up_fileNO);
+				System.out.println("noti_NO = " + noti_NO);
+
+				BoardNoticeImageVO boardNoticeImageVO = new BoardNoticeImageVO();
+				boardNoticeImageVO.setNoti_NO(Integer.parseInt(noti_NO));
+				boardNoticeImageVO.setUp_fileNO(Integer.parseInt(up_fileNO));
+				boardService.removeModImage(boardNoticeImageVO);
+				
+				File oldFile = new File(ARTICLE_IMAGE_REPO + "\\" + noti_NO + "\\" + up_fileName);
+				oldFile.delete();
+				
+				writer.print("success");
+			} catch (Exception e) {
+				writer.print("failed");
+			}
+
+		}
+		
+		// 수정 시 다중 이미지 업로드하기
+		private List<String> uploadModImageFile(MultipartHttpServletRequest multipartRequest) throws Exception {
+			List<String> fileList = new ArrayList<String>();
+			Iterator<String> fileNames = multipartRequest.getFileNames();
+			while (fileNames.hasNext()) {
+				String up_fileName = fileNames.next();
+				MultipartFile mFile = multipartRequest.getFile(up_fileName);
+				String originalFileName = mFile.getOriginalFilename();
+				if (originalFileName != "" && originalFileName != null) {
+					fileList.add(originalFileName);
+					File file = new File(ARTICLE_IMAGE_REPO + "\\" + up_fileName);
+					if (mFile.getSize() != 0) { // File Null Check
+						if (!file.exists()) { // 경로상에 파일이 존재하지 않을 경우
+							file.getParentFile().mkdirs(); // 경로에 해당하는 디렉토리들을 생성
+							mFile.transferTo(new File(ARTICLE_IMAGE_REPO + "\\" + "temp" + "\\" + originalFileName)); // 임시로
+						}
+					}
+				} else {
+					fileList.add(null);
+				}
+
+			}
+			return fileList;
+		}
 
 }
